simplify cli and library for benchmarking
